Relocate files; add new field into config.json; remove commands.
