+++ neo-cli/Settings.cs
@@ -52,11 +52,13 @@ namespace Neo
     {
         public string Path { get; }
         public bool ConsoleOutput { get; }
+        public bool Started { get; }
 
         public LoggerSettings(IConfigurationSection section)
         {
             this.Path = string.Format(section.GetSection("Path").Value, ProtocolSettings.Default.Magic.ToString("X8"));
             this.ConsoleOutput = section.GetSection("ConsoleOutput").Get<bool>();
+            this.Started = section.GetSection("Started").Get<bool>();
         }
     }
