+++ neo-cli/config.mainnet.json
@@ -1,5 +1,10 @@
 {
   "ApplicationConfiguration": {
+    "Logger": {
+      "Path": "SystemLogs_{0}",
+      "ConsoleOutput": false,
+      "Started": false
+    },
     "Storage": {
       "Engine": "LevelDBStore"
     },
